addressing review comments.
