+++ hadoop-ozone/integration-test-s3/src/test/java/org/apache/hadoop/ozone/s3/awssdk/OzoneS3SDKTests.java
@@ -20,7 +20,6 @@ package org.apache.hadoop.ozone.s3.awssdk;
 import org.apache.hadoop.ozone.MiniOzoneCluster;
 import org.apache.hadoop.ozone.s3.MultiS3GatewayService;
 import org.apache.hadoop.ozone.s3.awssdk.v1.AbstractS3SDKV1Tests;
-import org.apache.hadoop.ozone.s3.awssdk.v2.AbstractListDirectoryBucketsV2Tests;
 import org.apache.hadoop.ozone.s3.awssdk.v2.AbstractS3SDKV2Tests;
 import org.apache.ozone.test.ClusterForTests;
 import org.junit.jupiter.api.Nested;
@@ -50,11 +49,4 @@ abstract class OzoneS3SDKTests extends ClusterForTests<MiniOzoneCluster> {
     }
   }
 
-  @Nested
-  class ListDirectoryBuckets extends AbstractListDirectoryBucketsV2Tests {
-    @Override
-    public MiniOzoneCluster cluster() {
-      return getCluster();
-    }
-  }
 }
